\name{CRCAssigner-package}
\alias{CRCAssigner-package}
\alias{CRCAssigner}
\docType{package}
\title{
\packageTitle{CRCAssigner}
}
\description{
\packageDescription{CRCAssigner}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{CRCAssigner}
\packageIndices{CRCAssigner}
For predicting colorectal cancer subtypes (Sadanandam at al, 2013) using 786 gene signature or a reduced set of 38 genes.
}
\author{
\packageAuthor{CRCAssigner}

Maintainer: \packageMaintainer{CRCAssigner}
}
\references{
Sadanandam, A. et al. (2013) A colorectal cancer classification system that associates cellular phenotype and responses to therapy. Nature Medicine (19) 619-625. 
}

