\name{PAM38}
\alias{PAM38}
\docType{data}
\title{
PAM centroid for the 38 gene signature.
}
\description{
PAM centroid for the 38 gene signature.
}
\usage{data("PAM38")}
\format{
  A data frame with 38 observations on the following 5 variables.
  \describe{
    \item{\code{Enterocyte}}{a numeric vector}
    \item{\code{Goblet.like}}{a numeric vector}
    \item{\code{Inflammatory}}{a numeric vector}
    \item{\code{Stem.like}}{a numeric vector}
    \item{\code{TA}}{a numeric vector}
  }
}
\details{
PAM centroid for the 38 gene signature.
}
