\name{PAM786}
\alias{PAM786}
\docType{data}
\title{
PAM centroid for the 786 gene signature.
}
\description{
PAM centroid for the 786 gene signature.
}
\usage{data("PAM786")}
\format{
  A data frame with 786 observations on the following 5 variables.
  \describe{
    \item{\code{Inflammatory}}{a numeric vector}
    \item{\code{Goblet.like}}{a numeric vector}
    \item{\code{Enterocyte}}{a numeric vector}
    \item{\code{TA}}{a numeric vector}
    \item{\code{Stem.like}}{a numeric vector}
  }
}
\details{
PAM centroid for the 786 gene signature.
}
